projects/research-scout/requirements.txt
anthropic>=0.40.0
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
